Allow batching of commands
There may be cases where more than one command is to be send to the YKUSH
and the overhead of enumerating and opening the device each time is just
unecessary. Therefore add functions that allow arrays of commands to be
send and make the previous functions simply send arrays of size 1 to
prevent duplication of the command code.